refactor: Split the OCI implementation from PyOci.
The layer implementing the OCI specification was mixed with the PyOCI
specifics.

Now the implementation is split into 3 parts:
- HttpTransport: Sends (HTTP) requests to the upstream registry
- Oci: Implements the OCI specification
- PyOci: Implements the translation between Python packages and the OCI
  registry
